CCP-1559 - small refactor
